sql: Generate rowid by counter
The function `OP_NextIdEphemeral` do not produce unique ids.

The new way to get rowid is to create sequential a counter.
One of registers initializes with int64_t = 0 and increases after each
insert.

There are similar cases in `select.c` file, however, they are not as
straight-forward and would be fixed in future commits.

Related to #3297