sql: do DML in multiple steps instead of single step
For #2618 it is necessary to be able to get the last inserted or
updated tuple. But it is not possible on the current DML
implementation, which 1) regardless of count of changed
statements is done in a single sqlite3_step, 2) this sqlite3_step
does not return anything except SQLITE_DONE code.

Lets do DML in multiple steps, returning tuples one by one from
sqlite3_step by SQLITE_TUPLE code. The tuples are returned
excactly as **tuples**, not rows. It allows to avoid columns
copying and MessagePack building out of sqlite3_step.

Needed for #2618